requirements.txt
openai>=1.6.0
genanki>=0.13.0
python-dotenv>=1.0.0